move code around into src/compiler/optimize.rs
